use button ref's ownerDocument to set filter portal container
